correccionn de tipo de dato

+++ grpc-server/src/main/java/com/grpc/grpc_server/services/EventService.java
@@ -9,8 +9,8 @@ public interface EventService {
     List<Event> getAllEvents();
 
     Event createEvent(String nameEvent, String descriptionEvent, LocalDateTime dateTime, List<String> participantUsernames);
-    Event updateEvent(Long id, String nameEvent, String descriptionEvent, LocalDateTime dateTime, List<String> participantUsernames);
-    void deleteEvent(Long id);
-    void assignMember(Long eventId, String username);
-    void removeMember(Long eventId, String username);
+    Event updateEvent(Integer id, String nameEvent, String descriptionEvent, LocalDateTime dateTime, List<String> participantUsernames);
+    void deleteEvent(Integer id);
+    void assignMember(Integer eventId, String username);
+    void removeMember(Integer eventId, String username);
 }
